Put socket handling in a extra class
