+++ .gitignore
@@ -47,3 +47,6 @@ app.*.map.json
 /android/app/release
 /linux
 /windows
+
+# uml
+/uml
