+++ lib/src/fcp/fcp.dart
@@ -7,3 +7,5 @@ export 'package:free_chat/src/fcp/model/return_type.dart';
 export 'package:free_chat/src/fcp/model/persistence.dart';
 
 export 'package:free_chat/src/fcp/fcp_message_handler.dart';
+
+export 'package:free_chat/src/fcp/model/fcp_socket_handler.dart';
